Final commit
I had to debug the copy.js code, which I copied from wordcounter.js. I experimented with it and updated it using the Chrome Developer tool (F12). The code now counts characters in addition to words, and an estimated reading time has been included.

I added some updates to the wordcounter.html (a bulleted list for the instructions) and wordcounter.css (new dimensions to the IDs) files.